Deletetando um produto / Ajustes gerais

+++ lib/models/section.dart
@@ -96,7 +96,8 @@ class Section extends ChangeNotifier {
     }
 
     for (final original in originalItems) {
-      if (!items.contains(original)) {
+      if (!items.contains(original) &&
+          (original.image as String).contains('firebase')) {
         try {
           final ref = await storage.getReferenceFromUrl(
             original.image as String,
@@ -118,10 +119,12 @@ class Section extends ChangeNotifier {
     await firestoreRef.delete();
 
     for (final item in items) {
-      try {
-        final ref = await storage.getReferenceFromUrl(item.image as String);
-        await ref.delete();
-      } catch (_) {}
+      if ((item.image as String).contains('firebase')) {
+        try {
+          final ref = await storage.getReferenceFromUrl(item.image as String);
+          await ref.delete();
+        } catch (_) {}
+      }
     }
   }
